added stealing machanic

+++ src/Cargo.java
@@ -1,5 +1,5 @@
 public class Cargo extends Product {
-    Cargo(String _name, int _count) {
+    Cargo(String _name, long _count) {
         super(_name, _count);
     }
 }
